or1k:  Fix exception handling for fallthrough case
The exception handling code worked fine for cases when exception were
caught and we returned through __builtin_eh_return,  However when we
ended up ending up at end_of_stack in _Unwind_RaiseException we failed
to restore the link register and returned to nowhere.

Fixed this by always restoring the link register in the epilogue, and
overriding the saved stack link register in eh_return, so we can
restore from that.